Refactoring. Python package. Logging.
