+++ connectivity/config/logging_template.py
@@ -1,0 +1,31 @@
+import sys
+
+LOGGING_CONFIG = {
+    "version": 1,
+    "formatters": {
+        "standart": {"format": "%(asctime)s - %(levelname)s - %(message)s"},
+    },
+    "handlers": {
+        "console": {
+            "level": "INFO",
+            "formatter": "standart",
+            "class": "logging.StreamHandler",
+            "stream": sys.stdout,
+        },
+        "file": {
+            "level": "INFO",
+            "formatter": "standart",
+            "class": "logging.handlers.RotatingFileHandler",
+            "filename": "connectivity/logs/sensors-connectivity.log",
+            "maxBytes": 1024000,
+            "backupCount": 10,
+        },
+    },
+    "loggers": {
+        "sensors-connectivity": {
+            "handlers": ["console", "file"],
+            "level": "INFO",
+            "propagate": False,
+        },
+    },
+}
