--- launch/agent.launch
@@ -1,10 +1,0 @@
-<launch>
-	<arg name="config" default="$(find sensors_connectivity)/config/default.json" />
-
-    <node ns="connectivity" name="worker" pkg="sensors_connectivity" type="worker_node" output="screen" respawn="true">
-        <!--
-        <param name="model" value="QmQ7T25tg2DhmqQEdYcHJEDXRi4Mbu1JpoYMhceqDksjbc" />
-        -->
-        <param name="config" value="$(arg config)" />
-    </node>
-</launch>
