--- package.xml
@@ -1,23 +1,0 @@
-<?xml version="1.0"?>
-<package format="2">
-  <name>sensors_connectivity</name>
-  <version>0.8.0</version>
-  <description>Agent that offers data from sensors</description>
-  <author email="vadim.razorq@gmail.com">Vadim Manaenko</author>
-  <maintainer email="vadim.razorq@aira.life">Vadim Manaenko</maintainer>
-  <license>BSD3</license>
-
-  <buildtool_depend>catkin</buildtool_depend>
-  <build_depend>robonomics_msgs</build_depend>
-  <build_depend>robonomics_liability</build_depend>
-  <build_depend>rospy</build_depend>
-  <build_export_depend>robonomics_msgs</build_export_depend>
-  <build_export_depend>robonomics_liability</build_export_depend>
-  <build_export_depend>rospy</build_export_depend>
-  <exec_depend>robonomics_msgs</exec_depend>
-  <exec_depend>robonomics_liability</exec_depend>
-  <exec_depend>rospy</exec_depend>
-
-  <export>
-  </export>
-</package>
