+++ setup.py
@@ -1,12 +1,7 @@
-# -*- coding: utf-8 -*-
-## ! DO NOT MANUALLY INVOKE THIS setup.py, USE CATKIN INSTEAD
+import setuptools
 
-from distutils.core import setup
-from catkin_pkg.python_setup import generate_distutils_setup
-
-# fetch values from package.xml
-setup_args = generate_distutils_setup(
-    packages=["drivers", "stations", "feeders", "utils"],
-    package_dir={'': 'src', "utils": "utils"})
-setup(**setup_args)
-
+setuptools.setup(
+    name="sensors-connectivity",
+    version="0.10.0",
+    packages=["connectivity"],
+)
